Give several CDR tests a chance to finish
A few revisions back, these tests passed. Unfortunately, they passed because
originated channels were leaked, which meant that a 'core stop gracefully'
wouldn't actually stop Asterisk. On the plus side, this gave the CDR engine
a chance to log out the results; on the negative side, memory leaks.

Essentially we're so fast on killing these tests that on slow enough build
agents, there's a good chance Asterisk will unload the CDR CSV modules and be
shutting down before the test writes out the CDR entry. Making the origination
asynchronous helped, but not quite enough.

This patch lets the SimpleTestCase module delay shutting down. If we can't
get the CDRs written in a few seconds then, quite frankly, we deserve to fail.
And I'll have to try to figure out yet another solution to a racey problem.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4080 ad95026d-47f8-4470-b445-30a6b62014f3