* Made the masquerade super-test compatible with Asterisk 12.
* Made the test not have Park announce the parking space in case the
optimizations finish before the call has completed saying the parking
space.  The test may try to pickup the parked call before it has completed
parking it.  This is only going to occur if the local channel chain is
very short; such as a chain of 10.

(issue ASTERISK-22221)
Reported by: Matt Jordan


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4082 ad95026d-47f8-4470-b445-30a6b62014f3