Add a slight delay before calling the scenario start callbacks
In some cases when a test module's scenario start callback function is called
it performs an action before the SIPp process is ready. This is the case when
using PluggableModules.Originator to originate an outbound call to SIPp. This
simply adds a slight delay to allow time for the SIPp process to be ready
before calling the scenario start callbacks.



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4186 ad95026d-47f8-4470-b445-30a6b62014f3