testsuite: Add a test for reinvite to T.38 with directmedia
This is a test that sets up a SIP call from one SIPp endpoint to
another with directmedia with just audio and then a reinvite
is issued from one of the peers to the other to change the call
into a T.38 fax session. The test ensures that at this point,
media is routed back through Asterisk.

(closes issue ASTERISK-17273)
Reported by: dario
Review: https://reviewboard.asterisk.org/r/2854/


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4213 ad95026d-47f8-4470-b445-30a6b62014f3