Fix outbound dialing in app_page integration test
This fixes outbound dialing and address resolution in the app_page
integration test. The address "localhost" resolves to multiple
addresses and the first address in the list is "::1" which is now
accepted since IAX2 supports IPv6. In addition, chan_iax2 is explicitly
bound to 127.0.0.1 which can not accept IPv6 traffic. The outbound
dials now explicitly request "127.0.0.1".

Review: https://reviewboard.asterisk.org/r/2935/


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4265 ad95026d-47f8-4470-b445-30a6b62014f3