Lower the default debug and verbose levels.
For the testsuite, there's not a real need to have debug
and verbose levels super high. Most debug and verbose messages
are at level 5 or lower. If there is a test failure where a higher
debug or verbose level could be useful, then the test can be temporarily
adjusted to have a higher default level. As it stands, the high defaults
result in unnecessary cycles in the testsuite being wasted, as well as
unnecessarily large log files being created.

(Issue AST-1225)
reported by John Bigelow



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4271 ad95026d-47f8-4470-b445-30a6b62014f3