Don't require calltokens for the IAX2 hangupcause test
A test failure was recorded for the IAX2 hangupcause test on the 64-bit
build agent for CentOS 6. Upon inspecting the logs, it appears as if Asterisk
instance 2 rejected an IAX2 call from Asterisk instance 1 due to the call
token's timestamp in the inbound request having a time that was out of bounds
for what Asterisk instance 2 deemed valid. This was a sporadic failure that
is due in no small part to the general wonkiness that is that particular build
agent. This patch specifies that call tokens are not required for the users
involved in this test, as the timing sources on that build agent are not
exactly reliable.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4291 ad95026d-47f8-4470-b445-30a6b62014f3