Enable twisted logging
This patch enables logging from twisted to our Python log file. This can get
*quite* verbose, so hopefully this is only a temporary condition.

Note that we're enabling this do to some hung tests on the Bamboo build agents.
The hangs appear to be occurring due to an exception in one of the process
managers; however, that exception is not getting propagated up as it appears
to be getting caught down in twisted. This should allow us to figure out what
is throwing the exception the next time it happens.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4305 ad95026d-47f8-4470-b445-30a6b62014f3