show_subscriptions: Update test to not terminate on SIPp Scenario
The test object for the show_subscriptions test was inheriting from
the SIPpTestCase class. This class can only be used for tests that terminate
when the SIPp scenarios terminate, as the class will automatically stop the
reactor when the last SIPp scenario finishes. In the case of this test, the
SIPp scenario is merely used to create a subscription so that a subsequent
AMI action will successfully query for the subscriptions. Since the
scenario stops the reactor and kills Asterisk, this was unlikely to succeed
unless the AMI action ran at exactly the right time.

This patch changes the test object so that it inherits from TestCase. The
test object now manages the SIPp Scenario itself.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4408 ad95026d-47f8-4470-b445-30a6b62014f3