bridge_transfer_callee: Fix bouncing test
Channels entering the bridge in the bridge_transfer_callee test are in
a race to actually join the bridge, so the caller was not always
guaranteed to enter the bridge first. If the callee entered the bridge
first it would be hung up as the caller and the test would fail. The
test now correctly recognizes the caller and hangs up the channel when
appropriate.

(closes issue ASTERISK-22843)


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4428 ad95026d-47f8-4470-b445-30a6b62014f3