chanspy_w_mixmonitor: abandon Console channel driver; update to TestCase
This patch updates the chanspy_w_mixmonitor test such that it no longer uses
the console channel driver. It now uses a Local channel attached to a MoH
source to test the ChanSpy-ing. Additionally, it is now based on the TestCase
class, such that it gains the benefits of using that common base class.



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4462 ad95026d-47f8-4470-b445-30a6b62014f3