cdr/sqlite3: Add test for cdr_sqlite3_custom CDR backend
This patch adds a test for SQLite3 CDRs. It verifies that the records
are written out to a SQLite3 database, and that custom values can be written
out as well. It also verifies proper appending of userfield data, as well
as some other odds and ends of CDR behavior.

(issue ASTERISK-23162)

Review: https://reviewboard.asterisk.org/r/3152/


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4631 ad95026d-47f8-4470-b445-30a6b62014f3