dial tests: Fix tests for versions less than Asterisk 12
The tests modified by this commit made some improper assumptions about the
formatting and presence of the UserEvent events they were testing. It also
made som incorrect assumptions about the presence of pre-dial handlers,
which weren't added until version 11.

This patch fixes the assumptions such that the tests will run under the
oldest version of Asterisk possible for the test.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4668 ad95026d-47f8-4470-b445-30a6b62014f3