pjsip/subscribe/missing_aor: Fix race condition in starting SIPp scenario
The pjsip/subscribe/missing_aor has been failing consistently on some build
agents. The reason why is due to how the test was written. Each step in the
test is triggered on the success of the previous step, where the previous
step was typically an AMI action execution. However, the AMI actions for
updating external MWI occur when the Stasis message for modifying the
mailbox state has been created and dispatched, not when the core has
actually bothered to update the MWI. Hence, the test was running through
its various AMI actions and then sending the SIP SUBSCRIBE request before
all of the mailbox state had been updated.

On the fast build agent, this typically failed. On the slow build agent,
it typically succeeded due to Asterisk actually running faster than the
test python script.

This patch refactors the test to use the AMI event MessageWaiting to
trigger the next stage. The AMI event is raised when the core itself has
actually updated the mailbox, so the test knows for sure that the
mailboxes are in the correct state before it subscribes.

Review: https://reviewboard.asterisk.org/r/3217/


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4702 ad95026d-47f8-4470-b445-30a6b62014f3