originate-cdr-disposition: Fix test failure
This modifies the expected originate success condition for iterations 2 and 3.
In these iterations, the Dial operation completes with either a result of
"Congestion" or a result of "Failed". Since an Originate only succeeds if the
channel is Answered, it should be expected that these two operations would
fail as well.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4754 ad95026d-47f8-4470-b445-30a6b62014f3