Add a test for MWI state aggregation.
Tests that the mwi_aggregate option works as expected for PJSIP endpoints.
This is done by having two configured PJSIP endpoints: one that aggregates
MWI state and one that does not. Two mailboxes have their states updated. We
ensure that the mailbox counts reported to the endpoints are what we would
expect them to be.

Review: https://reviewboard.asterisk.org/r/3238



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4783 ad95026d-47f8-4470-b445-30a6b62014f3