testsuite: Update PJSIP tests that evaluate TOS settings to go back to decimal
Previously they were changed to strings because another patch changed them to
strings. Now they are back to decimals so I'm changing the expectations back
to decimals.  Also there was an issue with the disallow field (codecs) being
removed from sorcery objects and all of that data is now just in the allow
field. The disallow check is removed as well.

Review: https://reviewboard.asterisk.org/r/3323/


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4837 ad95026d-47f8-4470-b445-30a6b62014f3