rest_api/channels/originate: Check for hangups and end test accordingly
The rest_api/channels/originate test had no check for when it was done. The
reactor would eventually time out, and the test would fail since no one had
told it to succeed. We now look for the appropriate number of channel
destroyed events and - if we get them all - pass the test.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4846 ad95026d-47f8-4470-b445-30a6b62014f3