pcap_demo: skip test due to errors in yappcap/libpcap
The pcap libraries used by yappcap have a bug where they claim they cannot
allocate sufficient memory to being listening for packets. This causes the
pcap_demo test to fail some times with the following exception:

 File "yappcap.pyx", line 490, in yappcap.PcapLive.activate (yappcap.c:4333)
yappcap.PcapError: can't create rx ring on packet socket: Cannot allocate memory

The pcap generation capabilities are still useful in the test suite, but they
don't have to be used all of the time (and in general, are only used when we
need a packet capture from a test). Since this capability is only used when
needed, it doesn't really make sense to continue to have the continuous
integration tests fail due to a problem in yappcap.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4858 ad95026d-47f8-4470-b445-30a6b62014f3