apps/disa/nominal/authenticate: Fix conflict with extensions.lua
When pbx_lua is loaded with the default extensions.lua, it will add extension
1000@default. Depending on the order in which modules are loaded, this can
cause a 'match more' extension pattern match to believe that it should continue
to receive digits when looking for 100@default. Since the DISA test was
attempting to jump to 100@default, it would fail on the build agents, which
apparently load pbx_lua at a different time than my development machine. Go
Figure.

This patch simply makes the extension that we attempt to jump to 200@default,
which avoids any conflict with the sample extensions.lua extensions.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4973 ad95026d-47f8-4470-b445-30a6b62014f3