apps/disa/nominal/no_authentication: Add test for DISA that bypasses auth
This test spawns a Local channel that is put into the DISA application. The
application is launched with the 'no-password' option, such that the channel is
not prompted for a password. The channel enters in a valid extension, and the
test verifies that the channel is placed in that extension.

Review: https://reviewboard.asterisk.org/r/3460

ASTERISK-23522 #close
Reported by: Matt Jordan
patches:
  disa_no_authentication.diff uploaded by bford (License 6580)


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4976 ad95026d-47f8-4470-b445-30a6b62014f3