apps/disa/nominal/no_context: Add DISA test for not specifying a context
This test uses a Local channel to enter the DISA application. The application
is invoked without a context to jump to, relying instead on the default 'disa'
context. When the channel enters the correct extension, the channel is placed
into the extension in the 'disa' context. The test verifies that the channel
goes to the correct extension.

Review: https://reviewboard.asterisk.org/r/3459

ASTERISK-23523 #close
Reported by: Matt Jordan
patches:
  disa_no_context_diff uploaded by bford (License 6580)


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4977 ad95026d-47f8-4470-b445-30a6b62014f3