Increase AMI writetimeout for chatty tests
A number of tests generate a substantial amount of AMI events. When AMI debugging
is cranked up (which it now is if a test produces a 'full' log), the 100 ms write
timeout can be triggered in Asterisk and the test will fail. This patch adds some
manager.users.conf.inc files that bump up the writetimeout for those tests.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5040 ad95026d-47f8-4470-b445-30a6b62014f3