pbx/dialplan: Add an AMI user write timeout
Much like masquerade and a few other tests, the dialplan test is a bit chatty
with AMI user events. As such, given the amount of logging that occurs in
starpy, things can bog down and the connection get broken. This adds a user
level override to increase the write timeout.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5050 ad95026d-47f8-4470-b445-30a6b62014f3