bridge/parkcall_timeout: Update expected events for channel snapshot reduction
This patch updates tests in the Test Suite based on the code review at
https://reviewboard.asterisk.org/r/3568.

The only behaviour change that occurs as a result of r3568 are CDR results in
parkcall_timeout comebacktoorigin tests. The CDRs that changed are those that
occur when a channel is Parked. Previously, those records would contain the
context/extension of the location of the dial back when the park times out;
now, that information is only on the record that occurs when the dial is
made to the parker. This is actually more in line with what should happen: it
doesn't make much sense for information associated with a timeout callback to
be on the parking record.

Review: https://reviewboard.asterisk.org/r/3578/

ASTERISK-23811



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5119 ad95026d-47f8-4470-b445-30a6b62014f3