Add a library of pcap tools.
As you can see from the header of the file, Matt actually wrote
this, but I'm committing it. He's either being gracious or this
is some subversive attempt to get my name to show up in 'svn blame'
forever on this file, leading to finger-pointing and laughing in
my general direction.

It's probably the former, though.

Either way, this sort of thing is useful in general, but it will
be sorely needed for RLS tests to come. I'm placing this section
of code directly into trunk's testsuite though so that the changes
I actually make will be apparent when it comes time to do code
reviews for RLS stuff.



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5157 ad95026d-47f8-4470-b445-30a6b62014f3