testsuite: fail sipp test if sipp not installed
If sipp is not listed as a dependency and isn't
installed, the sipp test case hangs while it is
trying to kill a nonexistant process.  Instead,
check if it is installed and raise an exception
if not.  Also fixed the missing dependancy that
precipitated this discovery.

Review: https://reviewboard.asterisk.org/r/3884



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5341 ad95026d-47f8-4470-b445-30a6b62014f3