tests/manager: Add MixMonitor AMI test
This patch adds a test for the MixMonitor AMI action. It attaches a MixMonitor
to a Local channel and playback audio through the other side of the Local
channel. When the playback is complete, the resulting sound file is passed
through the SoundChecker pluggable module and verified to be of approximately
the correct size and that it has energy within it.

Review: https://reviewboard.asterisk.org/r/3795/

ASTERISK-24028 #close
patches:
  mixmonitor_record.diff uploaded by tcambron (License 6611)


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5441 ad95026d-47f8-4470-b445-30a6b62014f3