(rest_api/bridges|rest_api/channels)/playback: Add reverse/foward tests
This patch adds four tests for the rest_api/{bridges|channels}/playback tests:

(1) Test reversing a playback on a channel. This puts a local channel half into
    Stasis, and then plays back a sound on it (tt-monkeys). Upon starting the
    sound, the playback gets reversed using /control?operation=reverse and a
    TestEvent gets released that signifies a reversal has occurred.

(2) Test forwarding a playback on a channel. This does the same thing as Test 1
    except that it uses the /control?operation=forward ARI command to forward
    the playback, and looks for a FastForward TestEvent.

(3) Test reversing a playback on a bridge. This does the same thing as Test 1
    except that a channel gets put into a bridge, and the bridge instead is the
    one that starts the playback.

(4) Test forwarding a playback on a bridge. This does the same thing as Test 3
    except that it uses the /control?operation=forward command on the bridge.

Review: https://reviewboard.asterisk.org/r/3842/

ASTERISK-24092 #close
patches:
  playback.diff uploaded by cwolfe (License 6609)


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5448 ad95026d-47f8-4470-b445-30a6b62014f3