lib/python/asterisk/apptest: Fix callback errors with multiple actions
When an expected event has multiple actions, the current handling in apptest
will erroneously pass the next actions to execute to the callback incorrectly.
This is because it is passing the actions as a list, which is interpreted as
a list of arguments to the callback functions. This causes hilarity, as
a list of actions will only have some of those actions executed.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@5456 ad95026d-47f8-4470-b445-30a6b62014f3